Replace mathcolor with background-color
With background-color is easy to say of with element the mouse
is over.

The choice for orange as the background-color was because of a good
contrast with the white background and the back font.

Note: Look like that margin and padding doesn't effect mathml
elements. And border don't have a uniform behaviour.